climate_bot/requirements.txt
Django>=3.2
python-dotenv>=0.19.2
pyTelegramBotAPI>=4.5.1
requests>=2.26.0
APScheduler>=3.9.0
